Claim modal styles WIP.
